Create oop's
so i have started an oops project and my project idea has been fixed.
my idea is to create a payment system for our university which combines our college canteen for lunch, cafe for snacks and our store which offers us basic stationary. the idea is to create a web application or any sort of user accessible system and i have to create it in java. this app allows the students to skip long waiting queues to buy smthng by paying for things via online before hand and theyll be notified of the amount of what they want to purchase within the app which is updated each time stock is refilled, this allows shopkeepers who can also use the app to keep track of the demand and the product count. the system is based off of a coupon which is scanned to recieve the item. the online coupon is generated upon payment. i wanna do the same for my college canteen which would enable us to skip the long walking distance and hassle to go back to college, and instead have food from the college canteen from the money collectively paid as whole for hostel mess, a part of it is divided for lunch and based on booking maybe a day or even a week in advance we can have the same food in the canteen for the money thats already prepaid, bbut for day scholars they have to pay in real time to get their coupon. this would allow shopkeepers to keep track of demand and product count and students can track the amount of money they spend within the college and the efficient coupon system can avoid long standing queues allowing for faster purchase of products.